% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recordcount.R
\name{calculate_all_counts}
\alias{calculate_all_counts}
\title{calculate_all_counts}
\usage{
calculate_all_counts(unique_citations, citations, n_unique, db_colname)
}
\arguments{
\item{unique_citations}{A data frame representing unique citations.}

\item{citations}{A data frame representing citations.}

\item{n_unique}{A data frame with unique citations where the label of citation is "search".}

\item{db_colname}{A character string representing the name of the database column in the data frame.}
}
\value{
A data frame summarizing the counts and percentages, including totals for each metric.
The columns of the data frame are: 'Source', 'Records Imported', 'Distinct Records',
'Unique records', 'Non-unique Records', 'Source Contribution \%',
'Source Unique Contribution \%', 'Source Unique \%'.
}
\description{
This function takes in three data frames and a column name, and calculates various
counts related to citations from different sources. The counts include the number of
distinct records, imported records, unique records, non-unique records, and various percentages.
The function returns a data frame summarizing the counts and percentages,
including totals for each metric.
}
