% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recordcount.R
\name{calculate_counts}
\alias{calculate_counts}
\title{Calculate Counts of Records}
\usage{
calculate_counts(df1, df2, db_colname)
}
\arguments{
\item{df1}{A data frame representing unique citations. Must include a column matching \code{db_colname} which will be used for count calculations.}

\item{df2}{A data frame representing citations. Must include a column matching \code{db_colname} which will be used for count calculations.}

\item{db_colname}{The column name that exists in both \code{df1} and \code{df2} which will be used to calculate counts. Should be a character string.}
}
\value{
A data frame with the counts of records for each source in the input data frames. Includes columns for the source, records imported, distinct records, and a total row.
}
\description{
This function calculates the distinct record count and initial citations count for each source in the input data frames.
The function then combines these counts into one table, converts the necessary columns to numeric, adds a row for the totals, and returns the resulting data frame.
}
