% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recordcount.R
\name{calculate_counts_with_phase}
\alias{calculate_counts_with_phase}
\title{Calculate Counts with Phase}
\usage{
calculate_counts_with_phase(unique_citations, citations, db_colname)
}
\arguments{
\item{unique_citations}{A data frame of unique citations.}

\item{citations}{A data frame of all citations.}

\item{db_colname}{A string specifying the database column name in the citation data frames.}
}
\value{
A data frame that includes counts, precision, and recall calculated for each source.
}
\description{
This function calculates the counts, precision, and recall for each source in a dataset
of citations with respect to 'Screened' and 'Final' phases.
}
