% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{calculate_phase_count}
\alias{calculate_phase_count}
\title{Calculate phase counts, precision, and recall}
\usage{
calculate_phase_count(unique_citations, citations, db_colname)
}
\arguments{
\item{unique_citations}{A dataframe containing unique citations with phase information.}

\item{citations}{A dataframe containing all citations with phase information.}

\item{db_colname}{The name of the column representing the source database.}
}
\value{
A dataframe containing distinct counts, counts for different phases, precision,
and recall for each source, as well as totals.
}
\description{
This function calculates counts for different phases and calculates precision and recall
for each source based on unique citations and citations dataframe.
}
