% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{calculate_record_counts}
\alias{calculate_record_counts}
\title{Calculate record counts function
Calculate and combine counts of distinct records, imported records, and unique records for each database}
\usage{
calculate_record_counts(unique_citations, citations, n_unique, db_colname)
}
\arguments{
\item{unique_citations}{Dataframe. The dataframe for calculating distinct records count.}

\item{citations}{Dataframe. The dataframe for calculating records imported count.}

\item{n_unique}{Dataframe. The dataframe for calculating unique records count.}

\item{db_colname}{Character. The name of the column containing the database source information.}
}
\value{
A dataframe with counts of distinct records, imported records, and unique records for each source, including total counts and several calculated ratios and percentages.
}
\description{
This function calculates the counts of distinct records, records imported, and unique records for each database source.
It combines these counts into one dataframe and calculates several ratios and percentages related to the unique and distinct counts.
It also calculates the total for each count type.
}
