% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{calculate_source_counts}
\alias{calculate_source_counts}
\title{Calculate Source Counts}
\usage{
calculate_source_counts(unique_citations, citations, n_unique)
}
\arguments{
\item{unique_citations}{A dataframe of unique citations.}

\item{citations}{A dataframe of all citations.}

\item{n_unique}{A dataframe of unique citation counts.}
}
\value{
A dataframe of calculated citation counts.
}
\description{
This function processes citation data to generate a dataframe of citation counts.
}
\keyword{internal}
