% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{check_unique_search_meta}
\alias{check_unique_search_meta}
\title{Check whether each file contains results from unique database}
\usage{
check_unique_search_meta(files, ref_list)
}
\arguments{
\item{files}{List of file names}

\item{ref_list}{List of references}
}
\description{
If filenames are to be used to name databases, users should be aware if
one file contains results from more than one database.
}
\keyword{internal}
