% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{citation_summary_table}
\alias{citation_summary_table}
\title{Contribution summary table}
\usage{
citation_summary_table(
  citations,
  comparison_type = "sources",
  search_label = "search",
  screening_label = "final",
  top_n = NULL
)
}
\arguments{
\item{citations}{A deduplicated tibble as returned by \code{dedup_citations()}.}

\item{comparison_type}{Either "sources" to summarise and assess sources or "strings" to consider strings.}

\item{search_label}{One or multiple labels that identify initial search results (default: "search") - if multiple labels are provided, they are merged.}

\item{screening_label}{One or multiple label that identify screened records (default: "final") - if multiple are provided, each is compared to the search stage.}

\item{top_n}{Number of sources/strings to display, based on the number of total records they contributed at the search stage. Note that calculations and totals will still be based on all citations. Defaults to NULL, then all sources/strings are displayed.}
}
\value{
A tibble containing the contribution summary table, which shows the contribution of each source and the overall performance of the search
}
\description{
Create a summary table to show the contribution of each source and the overall performance of the search. For this to work,
labels need to be used that contrast a "search" stage with one or more later stages.
}
\examples{
# Load a sample data set
data("sample_citations")

# Deduplicate the citations
deduped_citations <- dedup_citations(sample_citations)

# Generate the citation summary table
summary_table <- citation_summary_table(deduped_citations)
}
