% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare_sources}
\alias{compare_sources}
\title{Compare duplicate citations across sources, labels, and strings}
\usage{
compare_sources(
  unique_data,
  comp_type = c("sources", "strings", "labels"),
  include_references = FALSE
)
}
\arguments{
\item{unique_data}{from ASySD, merged unique rows with duplicate IDs}

\item{comp_type}{Specify which fields are to be included. One or more of "sources", "strings" or "labels" - defaults to all.}

\item{include_references}{Should bibliographic detail be included in return?}
}
\value{
dataframe with indicators of where a citation appears, with sources/labels/strings as columns
}
\description{
Compare duplicate citations across sources, labels, and strings
}
\examples{
if (interactive()) {
  # Load example data from the package
  examplecitations_path <- system.file("extdata", "examplecitations.rds", package = "CiteSource")
  examplecitations <- readRDS(examplecitations_path)

  # Deduplicate citations and compare sources
  dedup_results <- dedup_citations(examplecitations)
  compare_sources(dedup_results, comp_type = "sources")
}
}
