% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R, R/recordcount.R
\name{count_sources}
\alias{count_sources}
\title{Count Sources}
\usage{
count_sources(df, db_colname)

count_sources(df, db_colname)
}
\arguments{
\item{df}{A data frame containing the raw data.}

\item{db_colname}{The name of the column in \code{df} that contains the database names.}
}
\value{
A data frame with columns 'Source' and 'count'.

A data frame of counts by source.
}
\description{
This function counts the unique citations per source.
}
\keyword{internal}
