% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recordcount.R
\name{create_combined_counts}
\alias{create_combined_counts}
\title{Create Combined Counts Function}
\usage{
create_combined_counts(citation_counts, unique_citations, citations)
}
\arguments{
\item{citation_counts}{A data frame returned by \code{extend_citation_counts()}.}

\item{citations}{A data frame of all citations.}

\item{#'}{unique_citations A data frame of unique citations.}
}
\value{
A data frame of combined citation counts.
}
\description{
This function combines the base citation counts with screening and final phases, and calculates precision/recall.
The totals computation includes the use of "⁶" and "⁷" superscripts which represent the counts of 'screened' and 'final' labels respectively in the citations data frame.
}
