% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recordcount.R
\name{extend_citation_counts}
\alias{extend_citation_counts}
\title{Extend Citation Counts Function}
\usage{
extend_citation_counts(citation_counts, unique_data)
}
\arguments{
\item{citation_counts}{A data frame returned by \code{count_sources()}.}

\item{unique_data}{A data frame of unique citations.}
}
\value{
A data frame of extended citation counts.
}
\description{
This function extends the base citation counts with unique/non-unique records and contribution metrics.
}
