% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{get_meta_vars}
\alias{get_meta_vars}
\title{Work out the metadata variable to be added to the internal reference}
\usage{
get_meta_vars(ref_list, meta_var)
}
\arguments{
\item{ref_list}{List of references.}

\item{meta_var}{The metadata variable.}
}
\description{
Work out the metadata variable to be added to the internal reference
}
\keyword{internal}
