% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dedup.R
\name{identify_true_matches}
\alias{identify_true_matches}
\title{This function identifies true pairs from matching pairs of citations and pairs which may be duplicates - for manual deduplication}
\usage{
identify_true_matches(pairs)
}
\arguments{
\item{pairs}{citation matches which may be duplicates}
}
\value{
Dataframe of true citation pairs
}
\description{
This function identifies true pairs from matching pairs of citations and pairs which may be duplicates - for manual deduplication
}
