% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_export_helpers.R
\name{parse_}
\alias{parse_}
\alias{parse_pubmed}
\alias{parse_ris}
\alias{parse_bibtex}
\alias{parse_csv}
\alias{parse_tsv}
\title{Parse bibliographic text in a variety of formats}
\usage{
parse_pubmed(x)

parse_ris(x, tag_naming = "best_guess")

parse_bibtex(x)

parse_csv(x)

parse_tsv(x)
}
\arguments{
\item{x}{A character vector containing bibliographic information in ris format.}

\item{tag_naming}{What format are ris tags in? Defaults to "best_guess" See \code{\link{read_refs}} for a list of accepted arguments.}
}
\value{
Returns an object of class \code{bibliography} (ris, bib, or pubmed formats) or \code{data.frame} (csv or tsv).
}
\description{
Text in standard formats - such as imported via \code{\link{readLines}} - can be parsed using a variety of standard formats. Use \code{\link{detect_parser}} to determine which is the most appropriate parser for your situation.
}
