% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{plot_source_counts}
\alias{plot_source_counts}
\title{Plot Source Counts}
\usage{
plot_source_counts(source_counts)
}
\arguments{
\item{source_counts}{A data frame produced by calculate_source_counts().}
}
\value{
A gt table.
}
\description{
This function takes the result of calculate_source_counts() and generates a formatted table using the gt package.
}
