% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{precision_sensitivity_table}
\alias{precision_sensitivity_table}
\title{precision_sensitivity_table}
\usage{
precision_sensitivity_table(data)
}
\arguments{
\item{data}{A data frame that must contain the columns "Source", "Distinct Records",
"Screened Included", "Final Included", "Precision", and "Recall". The "Source" column is used as the row names of the table.}
}
\value{
A gt object representing the table.
}
\description{
This function creates a table with footnotes for columns in the table.
It uses the gt package to create the table and adds footnotes to various columns.
}
