% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{read_citations}
\alias{read_citations}
\title{Import citations from file}
\usage{
read_citations(
  files = NULL,
  cite_sources = NULL,
  cite_strings = NULL,
  cite_labels = NULL,
  metadata = NULL,
  verbose = TRUE,
  tag_naming = "best_guess",
  only_key_fields = TRUE
)
}
\arguments{
\item{files}{One or multiple RIS or Bibtex files with citations.
Should be .bib or .ris files}

\item{cite_sources}{The origin of the citation files (e.g. "Scopus", "WOS", "Medline") - vector with one value per file, defaults to file names.}

\item{cite_strings}{Optional. The search string used (or another grouping to analyse) - vector with one value per file}

\item{cite_labels}{Optional. An additional label per file, for instance the stage of search - vector with one value per file}

\item{metadata}{A tibble with file names and metadata for each file. Can be specified as an \emph{alternative} to files, cite_sources, cite_strings and cite_labels.}

\item{verbose}{Should number of reference and allocation of labels be reported?}

\item{tag_naming}{Either a length-1 character stating how should ris tags be replaced (see details for a list of options), or an object inheriting from class \code{data.frame} containing user-defined replacement tags.}

\item{only_key_fields}{Should only key fields (e.g., those used by CiteCourse) be imported? If FALSE, all RIS data is retained. Can also be a character vector of field names to retain (after they have been renamed by the import function) in addition to the essential ones.}
}
\value{
A tibble with one row per citation
}
\description{
This function imports RIS and Bibtex files with citations and merges them
into one long tibble with one record per line.
}
\examples{
if (interactive()) {
  read_citations(c("res.ris", "res.bib"),
    cite_sources = c("CINAHL", "MEDLINE"),
    cite_strings = c("Search1", "Search2"),
    cite_labels = c("raw", "screened")
  )
  
  # or equivalently
  
  metadata_tbl <- tibble::tribble(
   ~files,     ~cite_sources, ~cite_strings, ~cite_labels, 
  "res.ris",  "CINAHL",      "Search1",     "raw", 
  "res.bib",  "MEDLINE",     "Search2",     "screened"
  )
  
  read_citations(metadata = metadata_tbl)
  
}
}
