% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{record_counts_table}
\alias{record_counts_table}
\title{record_counts_table}
\usage{
record_counts_table(data)
}
\arguments{
\item{data}{A data frame that must contain the columns "Source", "Records Imported",
and "Distinct Records". The "Source" column is used as the row names of the table.}
}
\value{
A gt object representing the table.
}
\description{
This function creates a table with footnotes for columns in the table.
It uses the gt package to create the table and adds footnotes to
the "Records Imported" and "Distinct Records" columns.
}
