% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShiny.R
\name{runShiny}
\alias{runShiny}
\title{A wrapper function to run Shiny Apps from \code{CiteSource}.}
\usage{
runShiny(app = "CiteSource")
}
\arguments{
\item{app}{Defaults to CiteSource - possibly other apps will be included in the future}
}
\value{
CiteSource shiny app
}
\description{
Running this function will launch the CiteSource shiny app
}
