% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_sources.R
\name{source_comparison}
\alias{source_comparison}
\title{Compare duplicate citations across sources}
\usage{
source_comparison(unique_data)
}
\arguments{
\item{unique_data}{from ASySD, merged unqiue rows with duplicate IDs}
}
\value{
dataframe with indicators of where a citation appears, with source as column
}
\description{
Compare duplicate citations across sources
}
