% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{calculate_phase_count}
\alias{calculate_phase_count}
\title{Calculate phase counts, precision, and recall}
\usage{
calculate_phase_count(unique_citations, citations, db_colname)
}
\arguments{
\item{unique_citations}{A dataframe containing unique citations with phase information.
The phase information must be provided in a column named 'cite_label' in the dataframe.}

\item{citations}{A dataframe containing all citations with phase information. The phase
information must be provided in a column named 'cite_label' in the dataframe.}

\item{db_colname}{The name of the column representing the source database.}
}
\value{
A dataframe containing distinct counts, counts for different phases, precision,
and recall for each source, as well as totals.
}
\description{
This function calculates counts for different phases and calculates precision and recall
for each source based on unique citations and citations dataframe. The phases should be
labeled as 'screened' and 'final' (case-insensitive) in the input dataframes. The function
will give a warning if these labels are not present in the input dataframes.
}
\details{
The function will give a warning if 'screened' and 'final' labels are not present
in the 'cite_label' column of the input dataframes.
}
\examples{
unique_citations <- data.frame(
db_source = c("Database1", "Database1", "Database2", "Database3", "Database3", "Database3"),
cite_label = c("screened", "final", "screened", "final", "screened", "final"),
duplicate_id = c(102, 102, 103, 103, 104, 104),
other_data = 1:6
)

citations <- data.frame(
db_source = c("Database1", "Database1", "Database1", "Database2", "Database2", "Database3"),
cite_label = c("screened", "final", "screened", "final", "screened", "final"),
other_data = 7:12
)

result <- calculate_phase_count(unique_citations, citations, "db_source")
result
}
