% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{calculate_phase_count}
\alias{calculate_phase_count}
\title{Calculate phase counts function
Calculate and combine counts of distinct records for each database and counts of 'Screened' and 'Final' labels}
\usage{
calculate_phase_count(unique_citations, citations, db_colname)
}
\arguments{
\item{unique_citations}{Dataframe. The dataframe for calculating distinct records count and also counts of 'Screened' and 'Final' labels.}

\item{citations}{Dataframe. Unused in the current implementation.}

\item{db_colname}{Character. The name of the column containing the database source information.}
}
\value{
A dataframe with counts of distinct records, 'Screened' and 'Final' labels for each source, including total counts, precision and recall.
}
\description{
This function calculates the counts of distinct records for each database source and the counts of 'Screened' and 'Final' labels.
It combines these counts into one dataframe, calculates precision and recall, and also calculates total counts for each count type.
}
