% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CiteSource.R
\docType{package}
\name{CiteSource-package}
\alias{CiteSource}
\alias{CiteSource-package}
\title{CiteSource: A package to compare sources of citation records}
\description{
The CiteSource package supports evidence aggregation by helping with the
processing of results of various searches in different sources. It allows to
deduplicate results while retaining meta-data on where those results were
found and then enables users to compare the contribution of different sources.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.eshackathon.org/CiteSource}
  \item Report bugs at \url{https://github.com/ESHackathon/CiteSource/issues}
}

}
\author{
\strong{Maintainer}: Trevor Riley \email{tnriley@gmail.com} (\href{https://orcid.org/0000-0002-6834-9802}{ORCID})

Authors:
\itemize{
  \item Trevor Riley \email{tnriley@gmail.com} (\href{https://orcid.org/0000-0002-6834-9802}{ORCID})
  \item Kaitlyn Hair \email{kaitlyn.hair@ed.ac.uk} (\href{https://orcid.org/0000-0003-0180-7343}{ORCID})
  \item Lukas Wallrich \email{lukas.wallrich@gmail.com} (\href{https://orcid.org/0000-0003-2121-5177}{ORCID})
  \item Matthew Grainger \email{matthewjamesgrainger@gmail.com} (\href{https://orcid.org/0000-0001-8426-6495}{ORCID})
  \item Sarah Young \email{sarahy@andrew.cmu.edu} (\href{https://orcid.org/0000-0002-8301-5106}{ORCID})
  \item Chris Pritchard \email{chris.pritchard@ntu.ac.uk} (\href{https://orcid.org/0000-0002-1143-9751}{ORCID})
  \item Neal Haddaway \email{nealhaddaway@gmail.com} (\href{https://orcid.org/0000-0003-3902-2234}{ORCID})
}

Other contributors:
\itemize{
  \item Martin Westgate (Author of included synthesisr fragments) [copyright holder]
  \item Eliza Grames (Author of included synthesisr fragments) [copyright holder]
  \item Kaitlyn Hair (Author of included ASySD deduplication code) [copyright holder]
  \item CAMARADES Group (Authors of ASySD (github.com/camaradesuk/ASySD)) [copyright holder]
}

}
