% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CiteSource.R
\docType{package}
\name{CiteSource}
\alias{CiteSource}
\title{CiteSource: A package to compare sources of citation records}
\description{
The CiteSource package supports evidence aggregation by helping with the
processing of results of various searches in different sources. It allows to
deduplicate results while retaining meta-data on where those results were
found and then enables users to compare the contribution of different sources
}
