% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare_sources}
\alias{compare_sources}
\title{Compare duplicate citations across sources, labels, and strings}
\usage{
compare_sources(
  unique_data,
  comp_type = c("sources", "strings", "labels"),
  include_references = FALSE
)
}
\arguments{
\item{unique_data}{from ASySD, merged unique rows with duplicate IDs}

\item{comp_type}{Specify which fields are to be included. One or more of "sources", "strings" or "labels" - defaults to all.}

\item{include_references}{Should bibliographic detail be included in return?}
}
\value{
dataframe with indicators of where a citation appears, with sources/labels/strings as columns
}
\description{
Compare duplicate citations across sources, labels, and strings
}
\examples{
unique_data <- ... # Load or create sample data here
result <- compare_sources(unique_data)
result_with_references <- compare_sources(unique_data, include_references = TRUE)
}
