% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_sources.R
\name{compare_sources}
\alias{compare_sources}
\title{Compare duplicate citations across sources}
\usage{
compare_sources(unique_data)
}
\arguments{
\item{unique_data}{from ASySD, merged unique rows with duplicate IDs}
}
\value{
dataframe with indicators of where a citation appears, with source as column
}
\description{
Compare duplicate citations across sources
}
