% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare_sources}
\alias{compare_sources}
\title{Compare duplicate citations across sources, labels, and strings}
\usage{
compare_sources(
  unique_data,
  comp_type = c("sources", "strings", "labels"),
  include_references = FALSE
)
}
\arguments{
\item{unique_data}{from ASySD, merged unique rows with duplicate IDs}

\item{comp_type}{Specify which fields are to be included. One or more of "sources", "strings" or "labels" - defaults to all.}

\item{include_references}{Should bibliographic detail be included in return?}
}
\value{
dataframe with indicators of where a citation appears, with sources/labels/strings as columns
}
\description{
Compare duplicate citations across sources, labels, and strings
}
\examples{
if (interactive()){
dedup_results <- dedup_citations(citations, merge_citations = TRUE)
unique_citations <- dedup_results$unique
compare_sources(unique_citations, comp_type = "sources")
}
}
