% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{count_sources}
\alias{count_sources}
\title{Count sources function
Count database source occurrences in a column}
\usage{
count_sources(df, db_colname)
}
\arguments{
\item{df}{Dataframe. The dataframe containing the data to be analyzed.}

\item{db_colname}{Character. The name of the column containing the database source information.}
}
\value{
A dataframe with the names of the sources and their corresponding counts.
}
\description{
This function counts the occurrences of different database sources present in a specific column of a dataframe.
It returns a dataframe with the counts for each source.
}
