% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{count_sources}
\alias{count_sources}
\title{Count Sources}
\usage{
count_sources(df, db_colname)
}
\arguments{
\item{df}{A data frame with the data to be processed.}

\item{db_colname}{A character string specifying the column name which contains the database source.}
}
\value{
A data frame with columns 'Source' and 'count'.
}
\description{
Count Sources
}
\keyword{internal}
