% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{count_unique}
\alias{count_unique}
\title{Count number of unique and non-unique citations from different sources, labels, and strings}
\usage{
count_unique(unique_data)
}
\arguments{
\item{unique_data}{from ASySD, merged unique rows with duplicate IDs}
}
\value{
dataframe with indicators of where a citation appears, with source/label/string as column
}
\description{
Count number of unique and non-unique citations from different sources, labels, and strings
}
