% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{count_unique}
\alias{count_unique}
\title{Count number of unique and non-unique citations from different sources, labels, and strings}
\usage{
count_unique(unique_data, include_references = FALSE)
}
\arguments{
\item{unique_data}{from ASySD, merged unique rows with duplicate IDs}

\item{include_references}{Should bibliographic detail be included in return?}
}
\value{
dataframe with indicators of where a citation appears, with source/label/string as column
}
\description{
Count number of unique and non-unique citations from different sources, labels, and strings
}
\examples{
# Load example data from the package
examplecitations_path <- system.file("extdata", "examplecitations.rds", package = "CiteSource")
examplecitations <- readRDS(examplecitations_path)

# Deduplicate citations
dedup_results <- dedup_citations(examplecitations)

# Count unique and non-unique citations
count_unique(unique_citations)
}
