% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dedup.R
\name{dedup_citations}
\alias{dedup_citations}
\title{Deduplicate citations - ASySD wrapper}
\usage{
dedup_citations(raw_citations)
}
\arguments{
\item{raw_citations}{Citation dataframe with relevant columns}
}
\value{
unique citations formatted for CiteSource
}
\description{
This function deduplicates citation data. Note that duplicates are assumed to published
in the same journal, so pre-prints and similar results will not be identified here.
}
\examples{
# Load example data from the package
examplecitations_path <- system.file("extdata", "examplecitations.rds", package = "CiteSource")
examplecitations <- readRDS(examplecitations_path)

# Deduplicate citations
dedup_results <- dedup_citations(examplecitations)

}
