% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dedup.R
\name{dedup_citations}
\alias{dedup_citations}
\title{Deduplicate citations - ASySD wrapper}
\usage{
dedup_citations(
  raw_citations,
  manual = FALSE,
  shiny_progress = FALSE,
  show_unknown_tags = FALSE
)
}
\arguments{
\item{raw_citations}{Citation dataframe with relevant columns}

\item{manual}{logical. If TRUE, manually specify pairs of duplicates to merge. Default is FALSE.}

\item{shiny_progress}{logical. If TRUE, show a progress bar in the Shiny app. Default is FALSE.}
}
\value{
unique citations formatted for CiteSource
}
\description{
This function deduplicates citation data. Note that duplicates are assumed to published
in the same journal, so pre-prints and similar results will not be identified here.
}
\examples{
# Load example data from the package
examplecitations_path <- system.file("extdata", "examplecitations.rds", package = "CiteSource")
examplecitations <- readRDS(examplecitations_path)

# Deduplicate citations
dedup_results <- dedup_citations(examplecitations)

}
