% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dedup.R
\name{dedup_citations}
\alias{dedup_citations}
\title{Deduplicate citations (ASySD Modified Function)}
\usage{
dedup_citations(
  raw_citations,
  manual_dedup = FALSE,
  merge_citations = FALSE,
  preferred_source = NULL
)
}
\arguments{
\item{raw_citations}{Citation dataframe with relevant columns}

\item{manual_dedup}{Logical value. Do you want to retrieve dataframe for manual deduplication?}

\item{merge_citations}{Logical value. Do you want to merge matching citations?}

\item{preferred_source}{citation source user wants to preferentially retain in dataset}
}
\value{
A list of 2 dataframes - unique citations and citations to be manually deduplicated if that option is selected
}
\description{
This function deduplicates citation data. Note that duplicates are assumed to published
in the same source, so pre-prints and similar results will not be identified here.
}
\examples{
# Load a sample citation dataset
data("citation_data")

# Deduplicate citations without manual deduplication and without merging matching citations
deduplicated_citations <- dedup_citations(citation_data, manual_dedup = FALSE, merge_citations = FALSE)

# Access the unique citations
unique_citations <- deduplicated_citations$unique

# Access the citations that require manual deduplication
manual_dedup_citations <- deduplicated_citations$manual_dedup
}
