% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dedup.R
\name{dedup_citations_add_manual}
\alias{dedup_citations_add_manual}
\title{Remove pairs with manual dedup - ASySD wrapper}
\usage{
dedup_citations_add_manual(unique_citations, additional_pairs)
}
\arguments{
\item{unique_citations}{Unique citations post deduplication}

\item{additional_pairs}{TRUE duplicate pairs}
}
\value{
unique citations formatted for CiteSource
}
\description{
This function deduplicates citation data. Note that duplicates are assumed to published
in the same journal, so pre-prints and similar results will not be identified here.
}
\examples{
# Load example data from the package
examplecitations_path <- system.file("extdata", "examplecitations.rds", package = "CiteSource")
examplecitations <- readRDS(examplecitations_path)

# Deduplicate citations
dedup_results <- dedup_citations(examplecitations)

}
