% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_bib}
\alias{export_bib}
\title{Export deduplicated citations to .bib file}
\usage{
export_bib(
  citations,
  filename = "citations.bib",
  include = c("sources", "labels", "strings")
)
}
\arguments{
\item{citations}{Dataframe with unique citations, resulting from \code{dedup_citations()}}

\item{filename}{Name (and path) of file, should end in .ris}

\item{include}{Character. One or more of sources, labels or strings}
}
\description{
This function saves deduplicated citations as a BibTex file with sources, labels and strings
included in the \code{note} field (if they were initially provided for any of the citations). Therefore,
beware that \strong{any \code{note} field that might be included in \code{citations} will be overwritten}. Also note that
\emph{existing files are overwritten without warning.}
}
\examples{
if (interactive()) {
dedup_results <- dedup_citations(citations, merge_citations = TRUE)
export_bib(dedup_results$unique, "cite_sources.bib", include = "sources")
}
}
