% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_csv}
\alias{export_csv}
\title{Export deduplicated citations with source data as csv file}
\usage{
export_csv(citations, filename = "citations.csv", separate = NULL)
}
\arguments{
\item{citations}{Dataframe with unique citations, resulting from \code{dedup_citations()}}

\item{filename}{Name (and path) of file, should end in .csv}

\item{separate}{Character vector indicating which (if any) of cite_source, cite_string and cite_label should be split into separate columns to faciliate further analysis.}
}
\description{
This function saves deduplicated citations as a csv file for further analysis and/or reporting.
Metadata can be separated into one column per source, label or string, which facilitates analysis.
}
\examples{
if (interactive()) {
dedup_results <- dedup_citations(citations, merge_citations = TRUE)
export_csv(dedup_results$unique, "cite_sources.csv", separate = "cite_source")
}
}
