% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_ris}
\alias{export_ris}
\title{Export data frame to RIS file}
\usage{
export_ris(
  citations,
  filename = "citations.ris",
  source_field = "DB",
  label_field = "C7",
  string_field = "C8"
)
}
\arguments{
\item{citations}{Dataframe to be exported to RIS file}

\item{filename}{Name (and path) of file, should end in .ris}

\item{source_field}{Field in \code{citations} representing the source. Default is "DB".}

\item{label_field}{Field in \code{citations} representing the label. Default is "C7".}

\item{string_field}{Field in \code{citations} representing additional string information. Default is "C8".}
}
\description{
This function saves a data frame as a RIS file with specified columns mapped to RIS fields. Note that
\emph{existing files are overwritten without warning.}
}
\examples{
if (interactive()) {
  # Load example data from the package
  examplecitations_path <- system.file("extdata", "examplecitations.rds", package = "CiteSource")
  examplecitations <- readRDS(examplecitations_path)
  dedup_results <- dedup_citations(examplecitations, merge_citations = TRUE)
  export_ris(dedup_results$unique, "cite_sources.ris", user_mapping = list("DB" = "cite_source_include", "C7" = "cite_label_include"))
}
}
