% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{precision_sensitivity_table}
\alias{precision_sensitivity_table}
\title{precision_sensitivity_table}
\usage{
precision_sensitivity_table(data)
}
\arguments{
\item{data}{A data.frame. The dataset to build the table from.
It should contain the columns 'screened', 'final', 'Precision', 'Recall'.}
}
\value{
A gt object representing the table.
}
\description{
This function creates a gt table from a given data, and
removes the 'screened' column and its associated footnotes if all its values are zero.
}
