% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{read_citations}
\alias{read_citations}
\title{Import citations from file}
\usage{
read_citations(
  files,
  cite_sources = NULL,
  cite_strings = NULL,
  cite_labels = NULL,
  tag_naming = "best_guess"
)
}
\arguments{
\item{files}{One or multiple RIS or Bibtex files with citations.
Should be .bib or .ris files}

\item{cite_sources}{The origin of the citation files (e.g. "Scopus", "WOS", "Medline") - vector with one value per file, defaults to file names.}

\item{cite_strings}{Optional. The search string used (or another grouping to analyse) - vector with one value per file}

\item{cite_labels}{Optional. An additional label per file, for instance the stage of search - vector with one value per file}

\item{tag_naming}{Either a length-1 character stating how should ris tags be replaced (see details for a list of options), or an object inheriting from class \code{data.frame} containing user-defined replacement tags.}
}
\value{
A tibble with one row per citation
}
\description{
This function imports RIS and Bibtex files with citations and merges them
into one long tibble with one record per line.
}
\examples{
\dontrun{
 read_citations(c("res.ris", "res.bib"),
 cite_sources = c("CINAHL", "MEDLINE"),
 cite_strings = c("Search1", "Search2"),
 cite_labels = c("raw", "screened"),
 }
}
