% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{record_counts}
\alias{record_counts}
\title{Record counts function
Calculate and combine counts of distinct records and imported records for each database}
\usage{
record_counts(unique_citations, citations, db_colname)
}
\arguments{
\item{unique_citations}{Dataframe. The dataframe for calculating distinct records count.}

\item{citations}{Dataframe. The dataframe for calculating records imported count.}

\item{db_colname}{Character. The name of the column containing the database source information.}
}
\value{
A dataframe with counts of distinct records and imported records for each source, including total counts.
}
\description{
This function calculates the counts of distinct records and records imported for each database source.
It combines these counts into one dataframe and calculates the total for each count type.
}
