% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{record_level_table}
\alias{record_level_table}
\title{Record-level table}
\usage{
record_level_table(citations, include = "source")
}
\arguments{
\item{citations}{A deduplicated tibble as returned by \code{dedup_citations()}.}

\item{include}{Which metadata should be included in the table? Defaults to 'source', can be replaced or expanded with 'label' and/or 'string'}
}
\value{
A tibble that can be turned into a pretty table with (e.g.) the \code{gt}-package
}
\description{
Creates a per-record table that shows which sources (and/or labels/strings) each item was found in.
}
