% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reimport.R
\name{reimport_csv}
\alias{reimport_csv}
\title{Reimport a CSV-file exported from CiteSource}
\usage{
reimport_csv(filename)
}
\arguments{
\item{filename}{Name (and path) of CSV file to be reimported, should end in .csv}
}
\value{
A data frame containing the imported citation data if all required columns are present.
}
\description{
This function reimports a csv file that was tagged and deduplicated by CiteSource.
It allows to continue with further analyses without repeating that step, and also
allows users to make any manual corrections to tagging or deduplication. Note that
this function only works on CSV files that were written with \code{export_csv(..., separate = NULL)}
}
\examples{
\dontrun{
#example usage
citations <- reimport_csv("path/to/citations.csv")
}

}
