% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShiny.R
\name{runShiny}
\alias{runShiny}
\title{A wrapper function to run Shiny Apps from \code{CiteSource}.}
\usage{
runShiny(app = "CiteSource", offer_install = interactive())
}
\arguments{
\item{app}{Defaults to CiteSource - possibly other apps will be included in the future}

\item{offer_install}{Should user be prompted to install required packages if they are missing?}
}
\value{
CiteSource shiny app
}
\description{
Running this function will launch the CiteSource shiny app
}
\examples{
\dontrun{
  # To run the CiteSource Shiny app:
  runShiny()
}
}
